move useFetchNativeCurrencyPrice out of balance dir
