+++ packages/hooks/src/balance/index.ts
@@ -1,3 +1,2 @@
 export { useBalance } from "./useBalance.js";
-export { useFetchNativeCurrencyPrice } from "./useFetchNativeCurrencyPrice.js";
 export { useWatchBalance } from "./useWatchBalance.js";
