+++ packages/hooks/src/balance/useBalance.ts
@@ -1,5 +1,5 @@
 import { formatEther, type Address, type Chain } from "viem";
-import { useFetchNativeCurrencyPrice } from "./useFetchNativeCurrencyPrice.js";
+import { useFetchNativeCurrencyPrice } from "../useFetchNativeCurrencyPrice.js";
 import { useCallback, useEffect, useState } from "react";
 import { useWatchBalance } from "./useWatchBalance.js";
