+++ packages/hooks/src/index.ts
@@ -1,2 +1,3 @@
 export { useAddress } from "./useAddress.js";
-export { useBalance, useFetchNativeCurrencyPrice, useWatchBalance } from "./balance/index.js";
+export { useBalance, useWatchBalance } from "./balance/index.js";
+export { useFetchNativeCurrencyPrice } from "./useFetchNativeCurrencyPrice.js";
